requirements.txt
streamlit>=1.28.0
pandas>=1.5.0
numpy>=1.21.0
scipy>=1.9.0
librosa>=0.9.0
pydub>=0.25.0
requests>=2.28.0
python-dotenv>=1.0.0
selenium>=4.15.0
vosk>=0.3.45
jellyfish>=1.2.0
sentence-transformers>=2.2.0
scikit-learn>=1.3.0
transformers>=4.36.0
# torch>=2.0.0  # Installed separately with CUDA support in Dockerfile
streamlit-autorefresh>=1.0.0
thefuzz>=0.19.0
openai-whisper>=20231117
psutil>=5.9.0
cryptography>=41.0.0
webdriver-manager>=4.0.0
bleach>=6.0.0
# llama-cpp-python>=0.2.0  # Installed separately with CUDA support in Dockerfile
accelerate>=0.24.0  # For optimized model loading
# Optional GPU optimizations (uncomment if needed):
# bitsandbytes>=0.41.0  # For quantization (optional optimization)
# flash-attn>=2.0.0  # For Flash Attention 2 (requires CUDA compilation, optional)
